refactor: refoctor hash join state
