+++ src/query/service/src/pipelines/pipeline_builder.rs
@@ -131,6 +131,8 @@ use crate::api::ExchangeInjector;
 use crate::pipelines::builders::build_append_data_pipeline;
 use crate::pipelines::builders::build_fill_missing_columns_pipeline;
 use crate::pipelines::processors::transforms::build_partition_bucket;
+use crate::pipelines::processors::transforms::hash_join::HashJoinBuildState;
+use crate::pipelines::processors::transforms::hash_join::HashJoinProbeState;
 use crate::pipelines::processors::transforms::hash_join::TransformHashJoinBuild;
 use crate::pipelines::processors::transforms::hash_join::TransformHashJoinProbe;
 use crate::pipelines::processors::transforms::range_join::TransformRangeJoinLeft;
@@ -151,7 +153,7 @@ use crate::pipelines::processors::transforms::TransformPartialAggregate;
 use crate::pipelines::processors::transforms::TransformPartialGroupBy;
 use crate::pipelines::processors::transforms::TransformWindow;
 use crate::pipelines::processors::AggregatorParams;
-use crate::pipelines::processors::JoinHashTable;
+use crate::pipelines::processors::HashJoinState;
 use crate::pipelines::processors::SinkRuntimeFilterSource;
 use crate::pipelines::processors::TransformCastSchema;
 use crate::pipelines::processors::TransformLimit;
@@ -168,7 +170,7 @@ pub struct PipelineBuilder {
     pub pipelines: Vec<Pipeline>,
 
     // Used in runtime filter source
-    pub join_state: Option<Arc<JoinHashTable>>,
+    pub join_state: Option<Arc<HashJoinBuildState>>,
     // record the index of join build side pipeline in `pipelines`
     pub index: Option<usize>,
 
@@ -729,13 +731,10 @@ impl PipelineBuilder {
         self.build_join_probe(join, state)
     }
 
-    fn build_join_state(&mut self, join: &HashJoin) -> Result<Arc<JoinHashTable>> {
-        JoinHashTable::create_join_state(
+    fn build_join_state(&mut self, join: &HashJoin) -> Result<Arc<HashJoinState>> {
+        HashJoinState::try_create(
             self.ctx.clone(),
-            &join.build_keys,
             join.build.output_schema()?,
-            join.probe.output_schema()?,
-            &join.probe_projections,
             &join.build_projections,
             HashJoinDesc::create(join)?,
         )
@@ -745,7 +744,7 @@ impl PipelineBuilder {
         &mut self,
         build: &PhysicalPlan,
         hash_join_plan: &HashJoin,
-        join_state: Arc<JoinHashTable>,
+        join_state: Arc<HashJoinState>,
     ) -> Result<()> {
         let build_side_context = QueryContext::create_from(self.ctx.clone());
         let mut build_side_builder = PipelineBuilder::create(
@@ -757,11 +756,14 @@ impl PipelineBuilder {
         let mut build_res = build_side_builder.finalize(build)?;
 
         assert!(build_res.main_pipeline.is_pulling_pipeline()?);
+        let build_state = HashJoinBuildState::try_create(
+            self.ctx.clone(),
+            &hash_join_plan.build_keys,
+            &hash_join_plan.build_projections,
+            join_state.clone(),
+        )?;
         let create_sink_processor = |input| {
-            let transform = TransformHashJoinBuild::create(
-                input,
-                TransformHashJoinBuild::attach(join_state.clone())?,
-            );
+            let transform = TransformHashJoinBuild::try_create(input, build_state.clone())?;
 
             if self.enable_profiling {
                 Ok(ProcessorPtr::create(ProcessorProfileWrapper::create(
@@ -775,7 +777,7 @@ impl PipelineBuilder {
         };
         if hash_join_plan.contain_runtime_filter {
             build_res.main_pipeline.duplicate(false)?;
-            self.join_state = Some(join_state);
+            self.join_state = Some(build_state);
             self.index = Some(self.pipelines.len());
         } else {
             build_res.main_pipeline.add_sink(create_sink_processor)?;
@@ -1647,17 +1649,25 @@ impl PipelineBuilder {
         )
     }
 
-    fn build_join_probe(&mut self, join: &HashJoin, state: Arc<JoinHashTable>) -> Result<()> {
+    fn build_join_probe(&mut self, join: &HashJoin, state: Arc<HashJoinState>) -> Result<()> {
         self.build_pipeline(&join.probe)?;
 
         let max_block_size = self.ctx.get_settings().get_max_block_size()? as usize;
         let func_ctx = self.ctx.get_function_context()?;
+
+        let probe_state = Arc::new(HashJoinProbeState::create(
+            self.ctx.clone(),
+            state.clone(),
+            &join.probe_projections,
+            join.probe.output_schema()?,
+            &join.join_type,
+        ));
         self.main_pipeline.add_transform(|input, output| {
             let transform = TransformHashJoinProbe::create(
                 input,
                 output,
                 join.projections.clone(),
-                TransformHashJoinProbe::attach(state.clone())?,
+                probe_state.clone(),
                 max_block_size,
                 func_ctx.clone(),
                 &join.join_type,
@@ -1844,6 +1854,7 @@ impl PipelineBuilder {
         let pipeline = &mut self.pipelines[self.index.unwrap()];
         let output_size = pipeline.output_len();
         debug_assert!(output_size % 2 == 0);
+
         let mut items = Vec::with_capacity(output_size);
         //           Join
         //          /   \
@@ -1855,10 +1866,10 @@ impl PipelineBuilder {
         for _ in 0..output_size / 2 {
             let input = InputPort::create();
             items.push(PipeItem::create(
-                ProcessorPtr::create(TransformHashJoinBuild::create(
+                ProcessorPtr::create(TransformHashJoinBuild::try_create(
                     input.clone(),
-                    TransformHashJoinBuild::attach(self.join_state.as_ref().unwrap().clone())?,
-                )),
+                    self.join_state.as_ref().unwrap().clone(),
+                )?),
                 vec![input],
                 vec![],
             ));
