+++ src/query/service/src/pipelines/processors/transforms/hash_join/mod.rs
@@ -14,9 +14,9 @@
 
 mod common;
 mod desc;
+mod hash_join_build_state;
+mod hash_join_probe_state;
 mod hash_join_state;
-mod hash_join_state_impl;
-mod join_hash_table;
 mod probe_join;
 mod probe_state;
 mod result_blocks;
@@ -26,9 +26,10 @@ mod transform_hash_join_probe;
 mod util;
 
 pub use desc::HashJoinDesc;
+pub use hash_join_build_state::HashJoinBuildState;
+pub use hash_join_probe_state::HashJoinProbeState;
+pub use hash_join_state::FixedKeyHashJoinHashTable;
 pub use hash_join_state::HashJoinState;
-pub use join_hash_table::FixedKeyHashJoinHashTable;
-pub use join_hash_table::JoinHashTable;
 pub use probe_state::ProbeState;
 pub use result_blocks::*;
 pub use transform_hash_join_build::TransformHashJoinBuild;
