+++ src/query/service/src/pipelines/processors/transforms/hash_join/probe_join/cross_join.rs
@@ -17,16 +17,16 @@ use common_expression::BlockEntry;
 use common_expression::DataBlock;
 use common_expression::Value;
 
+use crate::pipelines::processors::transforms::hash_join::HashJoinProbeState;
 use crate::pipelines::processors::transforms::hash_join::ProbeState;
-use crate::pipelines::processors::JoinHashTable;
 
-impl JoinHashTable {
-    pub(crate) fn probe_cross_join(
+impl HashJoinProbeState {
+    pub(crate) fn cross_join(
         &self,
         input: DataBlock,
         _probe_state: &mut ProbeState,
     ) -> Result<Vec<DataBlock>> {
-        let build_blocks = self.row_space.datablocks();
+        let build_blocks = self.hash_join_state.row_space.datablocks();
         let build_num_rows = build_blocks
             .iter()
             .fold(0, |acc, block| acc + block.num_rows());
