+++ src/query/service/src/pipelines/processors/transforms/hash_join/probe_join/left_mark.rs
@@ -27,10 +27,10 @@ use common_hashtable::HashJoinHashtableLike;
 use crate::pipelines::processors::transforms::hash_join::desc::MARKER_KIND_FALSE;
 use crate::pipelines::processors::transforms::hash_join::desc::MARKER_KIND_NULL;
 use crate::pipelines::processors::transforms::hash_join::desc::MARKER_KIND_TRUE;
+use crate::pipelines::processors::transforms::hash_join::HashJoinProbeState;
 use crate::pipelines::processors::transforms::hash_join::ProbeState;
-use crate::pipelines::processors::JoinHashTable;
 
-impl JoinHashTable {
+impl HashJoinProbeState {
     pub(crate) fn probe_left_mark_join<'a, H: HashJoinHashtableLike, IT>(
         &self,
         hash_table: &H,
@@ -49,7 +49,12 @@ impl JoinHashTable {
         let probe_column = input.get_by_offset(0).value.as_column().unwrap();
         // Check if there is any null in the probe column.
         if matches!(probe_column.validity().1, Some(x) if x.unset_bits() > 0) {
-            let mut has_null = self.hash_join_desc.marker_join_desc.has_null.write();
+            let mut has_null = self
+                .hash_join_state
+                .hash_join_desc
+                .marker_join_desc
+                .has_null
+                .write();
             *has_null = true;
         }
         let mut occupied = 0;
@@ -63,7 +68,7 @@ impl JoinHashTable {
             if (i & max_block_size) == 0 {
                 max_block_size <<= 1;
 
-                if self.interrupt.load(Ordering::Relaxed) {
+                if self.hash_join_state.interrupt.load(Ordering::Relaxed) {
                     return Err(ErrorCode::AbortedQuery(
                         "Aborted query, because the server is shutting down or the query was killed.",
                     ));
@@ -136,12 +141,22 @@ impl JoinHashTable {
         let probe_column = input.get_by_offset(0).value.as_column().unwrap();
         // Check if there is any null in the probe column.
         if matches!(probe_column.validity().1, Some(x) if x.unset_bits() > 0) {
-            let mut has_null = self.hash_join_desc.marker_join_desc.has_null.write();
+            let mut has_null = self
+                .hash_join_state
+                .hash_join_desc
+                .marker_join_desc
+                .has_null
+                .write();
             *has_null = true;
         }
 
         let _func_ctx = self.ctx.get_function_context()?;
-        let other_predicate = self.hash_join_desc.other_predicate.as_ref().unwrap();
+        let other_predicate = self
+            .hash_join_state
+            .hash_join_desc
+            .other_predicate
+            .as_ref()
+            .unwrap();
 
         let mut occupied = 0;
         let mut probe_indexes_len = 0;
@@ -149,7 +164,7 @@ impl JoinHashTable {
         let build_indexes = &mut probe_state.build_indexes;
         let build_indexes_ptr = build_indexes.as_mut_ptr();
 
-        let data_blocks = self.row_space.chunks.read();
+        let data_blocks = self.hash_join_state.row_space.chunks.read();
         let data_blocks = data_blocks
             .iter()
             .map(|c| &c.data_block)
@@ -157,14 +172,17 @@ impl JoinHashTable {
         let build_num_rows = data_blocks
             .iter()
             .fold(0, |acc, chunk| acc + chunk.num_rows());
-        let is_build_projected = self.is_build_projected.load(Ordering::Relaxed);
+        let is_build_projected = self
+            .hash_join_state
+            .is_build_projected
+            .load(Ordering::Relaxed);
 
         let mark_scan_map = unsafe { &mut *self.mark_scan_map.get() };
         let _mark_scan_map_lock = self.mark_scan_map_lock.lock();
 
         for (i, key) in keys_iter.enumerate() {
             let (mut match_count, mut incomplete_ptr) =
-                if self.hash_join_desc.from_correlated_subquery {
+                if self.hash_join_state.hash_join_desc.from_correlated_subquery {
                     hash_table.probe_hash_table(key, build_indexes_ptr, occupied, max_block_size)
                 } else {
                     self.probe_key(
@@ -186,7 +204,7 @@ impl JoinHashTable {
             probe_indexes_len += 1;
             if occupied >= max_block_size {
                 loop {
-                    if self.interrupt.load(Ordering::Relaxed) {
+                    if self.hash_join_state.interrupt.load(Ordering::Relaxed) {
                         return Err(ErrorCode::AbortedQuery(
                             "Aborted query, because the server is shutting down or the query was killed.",
                         ));
@@ -202,10 +220,11 @@ impl JoinHashTable {
                         None
                     };
                     let build_block = if is_build_projected {
-                        Some(
-                            self.row_space
-                                .gather(build_indexes, &data_blocks, &build_num_rows)?,
-                        )
+                        Some(self.hash_join_state.row_space.gather(
+                            build_indexes,
+                            &data_blocks,
+                            &build_num_rows,
+                        )?)
                     } else {
                         None
                     };
@@ -260,10 +279,15 @@ impl JoinHashTable {
             }
         }
 
-        if self.hash_join_desc.from_correlated_subquery {
+        if self.hash_join_state.hash_join_desc.from_correlated_subquery {
             // Must be correlated ANY subquery, we won't need to check `has_null` in `mark_join_blocks`.
             // In the following, if value is Null and Marker is False, we'll set the marker to Null
-            let mut has_null = self.hash_join_desc.marker_join_desc.has_null.write();
+            let mut has_null = self
+                .hash_join_state
+                .hash_join_desc
+                .marker_join_desc
+                .has_null
+                .write();
             *has_null = false;
         }
 
@@ -281,7 +305,7 @@ impl JoinHashTable {
             None
         };
         let build_block = if is_build_projected {
-            Some(self.row_space.gather(
+            Some(self.hash_join_state.row_space.gather(
                 &build_indexes[0..occupied],
                 &data_blocks,
                 &build_num_rows,
