+++ src/query/service/src/pipelines/processors/transforms/hash_join/probe_join/right_join.rs
@@ -24,11 +24,11 @@ use common_expression::DataBlock;
 use common_hashtable::HashJoinHashtableLike;
 use common_hashtable::RowPtr;
 
+use crate::pipelines::processors::transforms::hash_join::HashJoinProbeState;
 use crate::pipelines::processors::transforms::hash_join::ProbeState;
-use crate::pipelines::processors::JoinHashTable;
 use crate::sql::plans::JoinType;
 
-impl JoinHashTable {
+impl HashJoinProbeState {
     pub(crate) fn probe_right_join<'a, H: HashJoinHashtableLike, IT>(
         &self,
         hash_table: &H,
@@ -52,7 +52,7 @@ impl JoinHashTable {
         let mut result_blocks = vec![];
         let mut probe_indexes_len = 0;
 
-        let data_blocks = self.row_space.chunks.read();
+        let data_blocks = self.hash_join_state.row_space.chunks.read();
         let data_blocks = data_blocks
             .iter()
             .map(|c| &c.data_block)
@@ -60,18 +60,22 @@ impl JoinHashTable {
         let build_num_rows = data_blocks
             .iter()
             .fold(0, |acc, chunk| acc + chunk.num_rows());
-        let is_build_projected = self.is_build_projected.load(Ordering::Relaxed);
+        let is_build_projected = self
+            .hash_join_state
+            .is_build_projected
+            .load(Ordering::Relaxed);
         let outer_scan_map = unsafe { &mut *self.outer_scan_map.get() };
-        let right_single_scan_map = if self.hash_join_desc.join_type == JoinType::RightSingle {
-            outer_scan_map
-                .iter_mut()
-                .map(|sp| unsafe {
-                    std::mem::transmute::<*mut bool, *mut AtomicBool>(sp.as_mut_ptr())
-                })
-                .collect::<Vec<_>>()
-        } else {
-            vec![]
-        };
+        let right_single_scan_map =
+            if self.hash_join_state.hash_join_desc.join_type == JoinType::RightSingle {
+                outer_scan_map
+                    .iter_mut()
+                    .map(|sp| unsafe {
+                        std::mem::transmute::<*mut bool, *mut AtomicBool>(sp.as_mut_ptr())
+                    })
+                    .collect::<Vec<_>>()
+            } else {
+                vec![]
+            };
 
         for (i, key) in keys_iter.enumerate() {
             let (mut match_count, mut incomplete_ptr) = self.probe_key(
@@ -93,7 +97,7 @@ impl JoinHashTable {
                 loop {
                     // The matched_num must be equal to max_block_size.
                     debug_assert_eq!(matched_num, max_block_size);
-                    if self.interrupt.load(Ordering::Relaxed) {
+                    if self.hash_join_state.interrupt.load(Ordering::Relaxed) {
                         return Err(ErrorCode::AbortedQuery(
                             "Aborted query, because the server is shutting down or the query was killed.",
                         ));
@@ -110,14 +114,14 @@ impl JoinHashTable {
                         let nullable_columns = probe_block
                             .columns()
                             .iter()
-                            .map(|c| Self::set_validity(c, max_block_size, true_validity))
+                            .map(|c| self.set_validity(c, max_block_size, true_validity))
                             .collect::<Vec<_>>();
                         Some(DataBlock::new(nullable_columns, max_block_size))
                     } else {
                         None
                     };
                     let build_block = if is_build_projected {
-                        Some(self.row_space.gather(
+                        Some(self.hash_join_state.row_space.gather(
                             local_build_indexes,
                             &data_blocks,
                             &build_num_rows,
@@ -129,9 +133,16 @@ impl JoinHashTable {
                         self.merge_eq_block(probe_block, build_block, max_block_size);
 
                     if !result_block.is_empty() {
-                        if self.hash_join_desc.other_predicate.is_none() {
+                        if self
+                            .hash_join_state
+                            .hash_join_desc
+                            .other_predicate
+                            .is_none()
+                        {
                             result_blocks.push(result_block);
-                            if self.hash_join_desc.join_type == JoinType::RightSingle {
+                            if self.hash_join_state.hash_join_desc.join_type
+                                == JoinType::RightSingle
+                            {
                                 self.update_right_single_scan_map(
                                     local_build_indexes,
                                     &right_single_scan_map,
@@ -146,12 +157,18 @@ impl JoinHashTable {
                         } else {
                             let (bm, all_true, all_false) = self.get_other_filters(
                                 &result_block,
-                                self.hash_join_desc.other_predicate.as_ref().unwrap(),
+                                self.hash_join_state
+                                    .hash_join_desc
+                                    .other_predicate
+                                    .as_ref()
+                                    .unwrap(),
                             )?;
 
                             if all_true {
                                 result_blocks.push(result_block);
-                                if self.hash_join_desc.join_type == JoinType::RightSingle {
+                                if self.hash_join_state.hash_join_desc.join_type
+                                    == JoinType::RightSingle
+                                {
                                     self.update_right_single_scan_map(
                                         local_build_indexes,
                                         &right_single_scan_map,
@@ -166,7 +183,9 @@ impl JoinHashTable {
                             } else if !all_false {
                                 // Safe to unwrap.
                                 let validity = bm.unwrap();
-                                if self.hash_join_desc.join_type == JoinType::RightSingle {
+                                if self.hash_join_state.hash_join_desc.join_type
+                                    == JoinType::RightSingle
+                                {
                                     self.update_right_single_scan_map(
                                         local_build_indexes,
                                         &right_single_scan_map,
@@ -238,14 +257,14 @@ impl JoinHashTable {
             let nullable_columns = probe_block
                 .columns()
                 .iter()
-                .map(|c| Self::set_validity(c, probe_block.num_rows(), &validity))
+                .map(|c| self.set_validity(c, probe_block.num_rows(), &validity))
                 .collect::<Vec<_>>();
             Some(DataBlock::new(nullable_columns, validity.len()))
         } else {
             None
         };
         let build_block = if is_build_projected {
-            Some(self.row_space.gather(
+            Some(self.hash_join_state.row_space.gather(
                 &local_build_indexes[0..matched_num],
                 &data_blocks,
                 &build_num_rows,
@@ -256,9 +275,14 @@ impl JoinHashTable {
         let result_block = self.merge_eq_block(probe_block, build_block, matched_num);
 
         if !result_block.is_empty() {
-            if self.hash_join_desc.other_predicate.is_none() {
+            if self
+                .hash_join_state
+                .hash_join_desc
+                .other_predicate
+                .is_none()
+            {
                 result_blocks.push(result_block);
-                if self.hash_join_desc.join_type == JoinType::RightSingle {
+                if self.hash_join_state.hash_join_desc.join_type == JoinType::RightSingle {
                     self.update_right_single_scan_map(
                         &local_build_indexes[0..matched_num],
                         &right_single_scan_map,
@@ -273,12 +297,16 @@ impl JoinHashTable {
             } else {
                 let (bm, all_true, all_false) = self.get_other_filters(
                     &result_block,
-                    self.hash_join_desc.other_predicate.as_ref().unwrap(),
+                    self.hash_join_state
+                        .hash_join_desc
+                        .other_predicate
+                        .as_ref()
+                        .unwrap(),
                 )?;
 
                 if all_true {
                     result_blocks.push(result_block);
-                    if self.hash_join_desc.join_type == JoinType::RightSingle {
+                    if self.hash_join_state.hash_join_desc.join_type == JoinType::RightSingle {
                         self.update_right_single_scan_map(
                             &local_build_indexes[0..matched_num],
                             &right_single_scan_map,
@@ -293,7 +321,7 @@ impl JoinHashTable {
                 } else if !all_false {
                     // Safe to unwrap.
                     let validity = bm.unwrap();
-                    if self.hash_join_desc.join_type == JoinType::RightSingle {
+                    if self.hash_join_state.hash_join_desc.join_type == JoinType::RightSingle {
                         self.update_right_single_scan_map(
                             &local_build_indexes[0..matched_num],
                             &right_single_scan_map,
